refactor : Redis 역직렬화 및 커밋 안된 로그인 기능 포함 커밋
