+++ src/main/resources/application.properties
@@ -2,12 +2,6 @@ server.port=8080
 
 spring.profiles.active=local
 
-# \uB85C\uADF8 \uAD00\uB828 \uC124\uC815
-logging.level.jdbc.sqlonly=off
-logging.level.jdbc.sqltiming=info
-logging.level.jdbc.resultsettable=info
-logging.level.jdbc.audit=off
-logging.level.jdbc.resultset=off
-logging.level.jdbc.connection=off
-
-logging.level.org.hibernate.type: trace
+server.servlet.session.timeout=30m
+spring.session.redis.flush-mode=on_save
+spring.session.redis.namespace=spring:session:offcoupon
