+++ src/main/resources/data.sql
@@ -1,2 +1,2 @@
-INSERT INTO member (email, password, name, birthDate, phone, createdAt, updatedAt)
+INSERT INTO member (email, password, name, birthdate, phone, created_at, updated_at)
 VALUES('sejin@email.com','1234','박세진','1111-11-1122','010-222-222','2023-12-12','2023-12-12');
